+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Battery optimization status page row now links straight to the right dialog on OSS flavour (#1239, thanks @nycex)
 - Added Korean translations (#1244, thanks @whatareyoudoingfor)
-- Allow map rotation (#1236)
+- Allow map rotation (can be disabled in preferences) (#1236)
 - Add a scale bar to the OSM map view (#1263)
 - Allow scaling the OSM map tiles to help with map readability on some devices with the `osmTileScaleFactor` setting (#1223, #1262)
 - [OSS] Allow the user to request background location permission ("All the time") in the device settings (#1255)